+++ Utils/random_utils.h
@@ -7,7 +7,7 @@
 #include <random>
 using namespace std;
 
-int uniform_random_int(int max);
+int uniform_random_int(size_t max);
 double uniform_random_double(double max);
 int weighted_discrete_distribution(vector<double> weights);
 #endif //FASTER_RANDOM_UTILS_H
